weather/requirements.txt
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0annotated-types==0.7.0
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
Faker==37.0.1
mysqlclient==2.2.7
numpy==2.2.4
openai==1.66.5
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
python-decouple==3.8
requests==2.32.3